Adding a couple changes before I delete and reclone, because having issues with Conda env.
